dump_session() tests: merge test files using subprocess to test loading in a new session
